Ensure "dev" directory inside zap directory exists; delete dpkg check
This commit adds a check to ensure that inside the ceph-deploy user's
personal directory, a "dev" directory exists within the "zap" folder.
When passing devices to the zap function, it will create a file for
every device it has successfully zapped inside of "zap". Typically,
device names are specified using a "/dev/sda"-like syntax though, so
infact, the modules would try to create $HOME/zap/dev/$NAME for every
zapped disk. If the "dev" folder is missing, that will fail. Making
sure that the "dev" folder exists inside the "zap" folder solves the
problem.

In addition, this commit gets rid of the unless condition in the call
to install ceph. Running "ceph-deploy install" on a host where Ceph
already is installed doesn't cost a lot, but by having a hard-coded
dpkg call in here, we're in fact breaking compatibility with Red Hat
and CentOS systems for no real reason.